Replace BaseController validateId() method by a util
